pyproject.toml
[project]
name = "introspect"
version = "0.1.0"
description = "Toolkit for zero-fine-tuning introspection evaluation on open-weight LLMs."
requires-python = ">=3.10"
authors = [
  { name = "Introspection Research Team" }
]
readme = "README.md"
dependencies = [
  "torch>=2.2,<3.0",
  "transformers>=4.43,<5.0",
  "accelerate>=0.30,<1.0",
  "numpy>=1.26,<3.0",
  "pandas>=2.0,<3.0",
  "pyyaml>=6.0,<7.0",
  "tqdm>=4.66,<5.0",
  "matplotlib>=3.8,<4.0",
  "pytest>=8.2,<9.0",
  "ruff>=0.5,<0.6"
]

[project.scripts]
introspect-normalise-results = "introspect.src.tools.normalise_results:main"
introspect-run-benchmark = "introspect.src.benchmark.run_benchmark:main"

[project.optional-dependencies]
hydra = [
  "hydra-core>=1.3,<2.0"
]
train = [
  "trl>=0.9,<0.11",
  "peft>=0.11,<0.13",
  "datasets>=2.18,<3.0",
  "scikit-learn>=1.3,<2.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["introspect*"]
exclude = ["results*", "venv*"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
